Better structure for ai service
