+++ gripl/gripl-backend/src/main/kotlin/de/mertendieckmann/griplbackend/config/LlmConfig.kt
@@ -1,4 +1,4 @@
-package de.mertendieckmann.griplbackend.chat
+package de.mertendieckmann.griplbackend.config
 
 import dev.langchain4j.model.chat.ChatModel
 import dev.langchain4j.model.openai.OpenAiChatModel
